backend/requirements.txt
# Django Core (LTS Version - compatible with all packages)
Django==4.2.16
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS Headers
django-cors-headers==4.3.1

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.6.0

# HTTP Client for External API
requests==2.31.0

# Email & SMS
twilio==8.10.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Environment Variables
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.0

# Database Utilities
django-extensions==3.2.3

# Development Tools
django-debug-toolbar==4.2.0
